Agrego getUsers en RepoUsers
