+++ src/main/kotlin/com/grupox/wololo/model/repos/RepoUsers.kt
@@ -7,6 +7,7 @@ import java.util.*
 
 interface RepoUsers : MongoRepository<User, String> {
     fun findAllByIsAdminFalse(): List<User>                                     // gets all users that aren't admin
+    fun findAllByIsAdminFalseAndUsernameLike(username: String): List<User>      // gets all normal users that match
     fun findByIsAdminFalseAndMail(mail: String): Optional<User>                 // gets a user by mail only if it isn't admin
     fun findByIsAdminFalseAndId(id: ObjectId): Optional<User>                   // gets a user by id only if it isn´t admin
     fun findByIsAdminTrueAndId(id: ObjectId): Optional<User>                    // gets an admin by id
