+++ src/test/kotlin/com/grupox/wololo/MockitoHelper.kt
@@ -1,0 +1,12 @@
+package com.grupox.wololo
+
+import org.mockito.Mockito
+
+object MockitoHelper {
+    fun <T> anyObject(): T {
+        Mockito.any<T>()
+        return uninitialized()
+    }
+    @Suppress("UNCHECKED_CAST")
+    fun <T> uninitialized(): T =  null as T
+}
