+++ src/test/kotlin/com/grupox/wololo/controller_tests/GamesControllerTest.kt
@@ -24,15 +24,6 @@ import java.time.Duration
 import java.time.Instant
 import java.util.*
 
-object MockitoHelper {
-    fun <T> anyObject(): T {
-        Mockito.any<T>()
-        return uninitialized()
-    }
-    @Suppress("UNCHECKED_CAST")
-    fun <T> uninitialized(): T =  null as T
-}
-
 @SpringBootTest
 class GamesControllerTest {
     @Autowired
